requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.9.3=py39h17cfd9d_1
aiosignal=1.3.1=pyhd8ed1ab_0
annotated-types=0.6.0=pyhd8ed1ab_0
anyio=3.7.1=pyhd8ed1ab_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=23.2.0=pyh71513ae_0
aws-c-auth=0.7.16=h0d2f7a6_8
aws-c-cal=0.6.10=h677d54c_2
aws-c-common=0.9.14=h93a5062_0
aws-c-compression=0.2.18=h677d54c_2
aws-c-event-stream=0.4.2=h59ac3ca_6
aws-c-http=0.8.1=hfe5d766_7
aws-c-io=0.14.6=h9ac2cdb_1
aws-c-mqtt=0.10.3=hb8a1441_2
aws-c-s3=0.5.3=h4398043_0
aws-c-sdkutils=0.1.15=h677d54c_2
aws-checksums=0.1.18=h677d54c_2
aws-crt-cpp=0.26.4=h2547149_1
aws-sdk-cpp=1.11.267=ha761c4c_4
brotli-python=1.1.0=py39hb198ff7_1
bzip2=1.0.8=h93a5062_5
c-ares=1.27.0=h93a5062_0
ca-certificates=2024.2.2=hf0a4a13_0
certifi=2024.2.2=pyhd8ed1ab_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
datasets=2.18.0=pyhd8ed1ab_0
dill=0.3.8=pyhd8ed1ab_0
distro=1.9.0=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
fastapi=0.110.0=pyhd8ed1ab_0
filelock=3.13.1=pyhd8ed1ab_0
freetype=2.12.1=hadb7bae_2
frozenlist=1.4.1=py39h17cfd9d_0
fsspec=2024.2.0=pyhca7485f_0
gflags=2.2.2=hc88da5d_1004
glog=0.7.0=hc6770e3_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.4=pyhd8ed1ab_0
httpx=0.27.0=pyhd8ed1ab_0
huggingface_hub=0.21.4=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=73.2=hc8870d7_0
idna=3.6=pyhd8ed1ab_0
krb5=1.21.2=h92f50d5_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240116.1=cxx17_hebf3989_2
libarrow=15.0.2=h5e64418_1_cpu
libarrow-acero=15.0.2=hebf3989_1_cpu
libarrow-dataset=15.0.2=hebf3989_1_cpu
libarrow-flight=15.0.2=h1f98dca_1_cpu
libarrow-flight-sql=15.0.2=hb095944_1_cpu
libarrow-gandiva=15.0.2=h2c81988_1_cpu
libarrow-substrait=15.0.2=h50959cf_1_cpu
libblas=3.9.0=21_osxarm64_openblas
libbrotlicommon=1.1.0=hb547adb_1
libbrotlidec=1.1.0=hb547adb_1
libbrotlienc=1.1.0=hb547adb_1
libcblas=3.9.0=21_osxarm64_openblas
libcrc32c=1.1.2=hbdafb3b_0
libcurl=8.6.0=h2d989ff_0
libcxx=16.0.6=h4653b0c_0
libdeflate=1.20=h93a5062_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h93a5062_2
libevent=2.1.12=h2757513_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libgoogle-cloud=2.22.0=hbebe991_1
libgoogle-cloud-storage=2.22.0=h8a76758_1
libgrpc=1.62.1=h9c18a4f_0
libiconv=1.17=h0d3ecfb_2
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=21_osxarm64_openblas
libllvm16=16.0.6=haab561b_3
libnghttp2=1.58.0=ha4dd798_1
libopenblas=0.3.26=openmp_h6c19121_0
libparquet=15.0.2=h278d484_1_cpu
libpng=1.6.43=h091b4b1_0
libprotobuf=4.25.3=hbfab5d5_0
libre2-11=2023.09.01=h7b2c953_2
libsqlite=3.45.2=h091b4b1_0
libssh2=1.11.0=h7a5bd25_0
libthrift=0.19.0=h026a170_1
libtiff=4.6.0=h07db509_3
libutf8proc=2.8.0=h1a8c8d9_0
libwebp-base=1.3.2=hb547adb_0
libxcb=1.15=hf346824_0
libxml2=2.12.6=h0d0cfa8_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=18.1.2=hcd81f8e_0
lz4-c=1.9.4=hb7217d7_0
multidict=6.0.5=py39h02fc5c5_0
multiprocess=0.70.16=py39h17cfd9d_0
ncurses=6.4.20240210=h078ce10_0
numpy=1.26.4=py39h7aa2656_0
openai=1.14.2=pyhd8ed1ab_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.2.1=h0d3ecfb_1
orc=2.0.0=h3d3088e_0
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.1=py39h47e51b9_0
pillow=10.2.0=py39h755f0b7_0
pip=24.0=pyhd8ed1ab_0
pthread-stubs=0.4=h27ca646_1001
pyarrow=15.0.2=py39h8cf9275_1_cpu
pyarrow-hotfix=0.6=pyhd8ed1ab_0
pydantic=2.6.4=pyhd8ed1ab_0
pydantic-core=2.16.3=py39h8fec3ad_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.19=hd7ebdb9_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-dotenv=1.0.1=pyhd8ed1ab_0
python-multipart=0.0.9=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python-xxhash=3.4.1=py39h0f82c59_0
python_abi=3.9=4_cp39
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.1=py39h0f82c59_1
re2=2023.09.01=h4cba328_2
readline=8.2=h92ec313_1
regex=2023.12.25=py39h17cfd9d_0
requests=2.31.0=pyhd8ed1ab_0
safetensors=0.4.2=py39h8fec3ad_0
setuptools=69.2.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h17c5cce_0
sniffio=1.3.1=pyhd8ed1ab_0
starlette=0.36.3=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tokenizers=0.15.2=py39heb7857c_0
tqdm=4.66.2=pyhd8ed1ab_0
transformers=4.39.1=pyhd8ed1ab_0
typing-extensions=4.10.0=hd8ed1ab_0
typing_extensions=4.10.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
urllib3=2.2.1=pyhd8ed1ab_0
uvicorn=0.29.0=py39h2804cbe_0
wheel=0.43.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xxhash=0.8.2=hb547adb_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
yarl=1.9.4=py39h17cfd9d_0
zstd=1.5.5=h4f39d0f_0
